+++ src/commands/checkout.ts
@@ -1,8 +1,13 @@
 import * as vscode from 'vscode';
 import {TaskNode} from '../nodes';
+import {isWorkingTreeClean, checkoutToBranch} from '../api/git';
 
 const checkoutToTask = async (task: TaskNode, context: vscode.ExtensionContext) => {
-    console.log('About to checkout to ', task.data.branchName);
+    if (!isWorkingTreeClean) {
+        return vscode.window.showErrorMessage('Please, commit all your cahnges before doing the checkout.');
+    }
+
+    checkoutToBranch(task.data.branchName);
 };
 
 export default checkoutToTask;