+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Change Log
-All notable changes to the "novum-webapp-workplace" extension will be documented in this file.
+
+All notable changes to the "webapp-workplace" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
-- Initial release
+
+-   Initial release